# api-spec/v1/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: PagoPA Ecommerce Watchdog Service API
  description: API service to monitor deadletter transactions and manually track recovery or follow-up actions.
servers:
  - url: https://${host}
tags:
  - name: deadletter-transactions
    description: Deadletter transaction management
  - name: deadletter-transactions-actions
    description: Manual actions or notes related to deadletter transactions
  - name: authentication
    description: User authentication
  - name: deadletter-transaction-notes
    description: Deadletter transaction notes management
paths:
  /deadletter-transactions:
    get:
      tags:
        - deadletter-transactions
      summary: List all deadletter transactions
      operationId: listDeadletterTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: '2025-08-19'
          description: Filter transactions by date (format YYYY-MM-DD)
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Searched page number, starting from 0
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 10
          description: Max element per page
      responses:
        '200':
          description: List of deadletter transactions with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  deadletterTransactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeadletterTransaction'
                  page:
                    $ref: '#/components/schemas/PageInfo'
                required:
                  - deadletterTransactions
                  - page
        '400':
          description: Invalid request or query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /deadletter-transactions/{transactionId}/actions:
    get:
      tags:
        - deadletter-transactions-actions
      summary: List actions for a specific deadletter transaction
      operationId: listActionsForDeadletterTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Deadletter transaction unique ID
      responses:
        '200':
          description: List of actions taken on the specified deadletter transaction
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeadletterTransactionAction'
        '400':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
    post:
      tags:
        - deadletter-transactions-actions
      summary: Add a new action to a deadletter transaction
      operationId: addActionToDeadletterTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Deadletter transaction unique ID
      requestBody:
        description: Action to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeadletterTransactionActionInput'
      responses:
        '201':
          description: Action added successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /deadletter-transactions/actions/bulk:
    post:
      tags:
        - deadletter-transactions-actions
      summary: Add a new action to multiple deadletter transactions
      operationId: addActionToDeadletterTransactions
      security:
        - bearerAuth: []
      requestBody:
        description: Action to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeadletterTransactionsActionInput'
      responses:
        '201':
          description: Action added successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /deadletter-transactions/{transactionId}/notes:
    post:
      tags:
        - deadletter-transaction-notes
      summary: Add a new note to a deadletter transaction
      operationId: addNoteToDeadletterTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Deadletter transaction unique ID
      requestBody:
        description: The text of the note
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '201':
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '422':
          description: Unprocessable entity. Number of notes already equal to the limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /deadletter-transactions/{transactionId}/notes/{noteId}:
    put:
      tags:
        - deadletter-transaction-notes
      summary: Modify an existing note connected to a deadletter transaction
      operationId: updateNoteDeadletterTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Deadletter transaction unique ID
        - name: noteId
          in: path
          required: true
          schema:
            type: string
          description: Note unique ID
      requestBody:
        description: The text of the note
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '204':
          description: Note modified successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction or note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
    delete:
      tags:
        - deadletter-transaction-notes
      summary: Delete an existing note connected to a deadletter transaction
      operationId: deleteNoteDeadletterTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Deadletter transaction unique ID
        - name: noteId
          in: path
          required: true
          schema:
            type: string
          description: Note unique ID
      responses:
        '204':
          description: Note deleted successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '404':
          description: Deadletter transaction or note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /deadletter-transactions/notes:
    post:
      tags:
        - deadletter-transaction-notes
      summary: Get all notes grouped by given transactionIds
      operationId: getNotesByTransactionIdList
      security:
        - bearerAuth: []
      requestBody:
        description: List of transactionIds
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesRequest'
      responses:
        '200':
          description: List of notes grouped by transactionIds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /authenticate:
    post:
      tags:
        - authentication
      summary: User authentication
      operationId: authenticateUser
      requestBody:
        description: Body containing username and password
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationCredentials"
      responses:
        '200':
          description: Successful authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationOk'
        '400':
          description: Bad Request. The request body is missing required fields or is malformed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '401':
          description: Unauthorized. The provided credentials are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /actions:
    get:
      tags:
        - action-types
      summary: List of available actions
      operationId: listActions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of available actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionType'
        '401':
          description: Unauthorized. Missing or invalid authentication token.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PageInfo:
      description: Information about the returned query page
      type: object
      properties:
        current:
          type: integer
          description: Current returned page index (0-based)
        total:
          type: integer
          description: Total pages for the query (based on requested page size)
        results:
          type: integer
          description: Transactions returned into the current page
      required:
        - current
        - results
        - total
    DeadletterTransaction:
      type: object
      properties:
        transactionId:
          type: string
        insertionDate:
          type: string
          format: date-time
        paymentToken:
          type: string
        paymentMethodName:
          type: string
        pspId:
          type: string
        eCommerceStatus:
          type: string
        gatewayAuthorizationStatus:
          type: string
        paymentEndToEndId:
          type: string
        operationId:
          type: string
        deadletterTransactionDetails:
          type: object
        eCommerceDetails:
          type: object
        nodoDetails:
          type: object
        npgDetails:
          type: object
      required:
        - id
        - insertionDate
        - transactionId
        - paymentToken
        - paymentMethodName
        - pspId
        - eCommerceStatus
        - deadletterTransactionDetails
    DeadletterTransactionAction:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
        userId:
          type: string
        action:
          $ref: '#/components/schemas/ActionType'
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - transactionId
        - userId
        - action
        - timestamp
    ActionType:
      type: object
      properties:
        value:
          type: string
        type:
          type: string
          enum:
            - "FINAL"
            - "NOT_FINAL"
      required:
        - value
        - type
    Note:
      type: object
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 300
        noteId:
          type: string
        transactionId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
      required:
        - note
        - noteId
        - transactionId
        - createdAt
        - updatedAt
        - userId
    NoteInput:
      type: object
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 300
      required:
        - note
    NotesRequest:
      type: object
      properties:
        transactionIds:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
      required:
        - transactionIds
    TransactionNotes:
      type: object
      properties:
        transactionId:
          type: string
        notesList:
          type: array
          items:
            $ref: '#/components/schemas/Note'
      required:
        - transactionId
        - notesList
    NotesResponse:
      type: array
      items:
        $ref: '#/components/schemas/TransactionNotes'
    DeadletterTransactionActionInput:
      type: object
      properties:
        value:
          type: string
      required:
        - value
    DeadletterTransactionsActionInput:
      type: object
      properties:
        transactionIds:
          type: array
          items:
            type: string
        value:
          type: string
      required:
        - value
    AuthenticationCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required:
        - username
        - password
    AuthenticationOk:
      type: object
      properties:
        urlRedirect:
          type: string
          example: https://watchdog-fe/#token=123
      required:
        - urlRedirect
    ProblemJson:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://example.com/problem/invalid-request
        title:
          type: string
          example: Invalid Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: The provided input is not valid
        instance:
          type: string
          format: uri
          example: /deadletter-transactions/123/actions